# packages/contracts/openapi.json
{
  "components": {
    "schemas": {
      "AccountConnections": {
        "properties": {
          "contacts": {
            "items": {
              "$ref": "#/components/schemas/VerifiedContact"
            },
            "title": "Contacts",
            "type": "array"
          },
          "identities": {
            "items": {
              "$ref": "#/components/schemas/LinkedIdentity"
            },
            "title": "Identities",
            "type": "array"
          }
        },
        "title": "AccountConnections",
        "type": "object"
      },
...
      "ApiResponse_AccountConnections_": {
        "properties": {
          "code": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "string"
              }
            ],
            "title": "Code"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AccountConnections"
              },
              {
                "type": "null"
              }
            ]
          },
          "msg": {
            "title": "Msg",
            "type": "string"
          }
        },
        "required": ["code", "msg"],
        "title": "ApiResponse[AccountConnections]",
        "type": "object"
      },
...
      "ApiResponse_IdentitySession_": {
        "properties": {
          "code": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "string"
              }
            ],
            "title": "Code"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/IdentitySession"
              },
              {
                "type": "null"
              }
            ]
          },
          "msg": {
            "title": "Msg",
            "type": "string"
          }
        },
        "required": ["code", "msg"],
        "title": "ApiResponse[IdentitySession]",
        "type": "object"
      },
...
      "ApiResponse_OAuthStart_": {
        "properties": {
          "code": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "string"
              }
            ],
            "title": "Code"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OAuthStart"
              },
              {
                "type": "null"
              }
            ]
          },
          "msg": {
            "title": "Msg",
            "type": "string"
          }
        },
        "required": ["code", "msg"],
        "title": "ApiResponse[OAuthStart]",
        "type": "object"
      },
...
      "ApiResponse_VerificationChallenge_": {
        "properties": {
          "code": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "string"
              }
            ],
            "title": "Code"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/VerificationChallenge"
              },
              {
                "type": "null"
              }
            ]
          },
          "msg": {
            "title": "Msg",
            "type": "string"
          }
        },
        "required": ["code", "msg"],
        "title": "ApiResponse[VerificationChallenge]",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "IdentitySession": {
        "properties": {
          "access_token_expire": {
            "format": "date-time",
            "title": "Access Token Expire",
            "type": "string"
          },
          "account_created": {
            "title": "Account Created",
            "type": "boolean"
          },
          "linked_by": {
            "enum": [
              "identity",
              "verified_phone",
              "verified_email",
              "new_account"
            ],
            "title": "Linked By",
            "type": "string"
          },
          "provider": {
            "title": "Provider",
            "type": "string"
          },
          "refresh_token_expire": {
            "format": "date-time",
            "title": "Refresh Token Expire",
            "type": "string"
          },
          "user_id": {
            "title": "User Id",
            "type": "integer"
          }
        },
        "required": [
          "user_id",
          "account_created",
          "linked_by",
          "provider",
          "access_token_expire",
          "refresh_token_expire"
        ],
        "title": "IdentitySession",
        "type": "object"
      },
...
      "OAuthStart": {
        "properties": {
          "authorization_url": {
            "title": "Authorization Url",
            "type": "string"
          },
          "flow": {
            "default": "login",
            "enum": ["login", "connect"],
            "title": "Flow",
            "type": "string"
          },
          "provider": {
            "enum": ["google", "qq", "wechat"],
            "title": "Provider",
            "type": "string"
          }
        },
        "required": ["provider", "authorization_url"],
        "title": "OAuthStart",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": ["loc", "msg", "type"],
        "title": "ValidationError",
        "type": "object"
      },
      "VerificationChallenge": {
        "properties": {
          "challenge_id": {
            "title": "Challenge Id",
            "type": "string"
          },
          "channel": {
            "enum": ["phone", "email"],
            "title": "Channel",
            "type": "string"
          },
          "expires_in": {
            "title": "Expires In",
            "type": "integer"
          },
          "resend_after": {
            "title": "Resend After",
            "type": "integer"
          },
          "target_hint": {
            "title": "Target Hint",
            "type": "string"
          }
        },
        "required": [
          "challenge_id",
          "channel",
          "target_hint",
          "expires_in",
          "resend_after"
        ],
        "title": "VerificationChallenge",
        "type": "object"
      },
      "VerificationCompleteRequest": {
        "properties": {
          "challenge_id": {
            "maxLength": 160,
            "minLength": 20,
            "title": "Challenge Id",
            "type": "string"
          },
          "code": {
            "pattern": "^\\d{6}$",
            "title": "Code",
            "type": "string"
          }
        },
        "required": ["challenge_id", "code"],
        "title": "VerificationCompleteRequest",
        "type": "object"
      },
      "VerificationRequest": {
        "properties": {
          "channel": {
            "enum": ["phone", "email"],
            "title": "Channel",
            "type": "string"
          },
          "target": {
            "maxLength": 320,
            "minLength": 3,
            "title": "Target",
            "type": "string"
          }
        },
        "required": ["channel", "target"],
        "title": "VerificationRequest",
        "type": "object"
      },
      "VerifiedContact": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "enum": ["phone", "email"],
            "title": "Kind",
            "type": "string"
          },
          "value_hint": {
            "title": "Value Hint",
            "type": "string"
          },
          "verification_source": {
            "title": "Verification Source",
            "type": "string"
          },
          "verified_at": {
            "format": "date-time",
            "title": "Verified At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "kind",
          "value_hint",
          "verification_source",
          "verified_at"
        ],
        "title": "VerifiedContact",
        "type": "object"
      }
    },
    "securitySchemes": {
      "CookieSession": {
        "description": "HttpOnly access-token cookie issued by Sun World authentication endpoints.",
        "in": "cookie",
        "name": "access_token",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "Sun World API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/auth/connections": {
      "get": {
        "operationId": "list_account_connections_auth_connections_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_AccountConnections_"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "CookieSession": []
          }
        ],
        "summary": "List Account Connections",
        "tags": ["identity"]
      }
    },
    "/auth/connections/verification/complete": {
      "post": {
        "operationId": "complete_connection_verification_auth_connections_verification_complete_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerificationCompleteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_AccountConnections_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "CookieSession": []
          }
        ],
        "summary": "Complete Connection Verification",
        "tags": ["identity"]
      }
    },
    "/auth/connections/verification/request": {
      "post": {
        "operationId": "request_connection_verification_auth_connections_verification_request_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerificationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_VerificationChallenge_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "CookieSession": []
          }
        ],
        "summary": "Request Connection Verification",
        "tags": ["identity"]
      }
    },
...
    "/auth/oauth/{provider}/callback": {
      "get": {
        "operationId": "complete_oauth_callback_auth_oauth__provider__callback_get",
        "parameters": [
          {
            "in": "path",
            "name": "provider",
            "required": true,
            "schema": {
              "title": "Provider",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "state",
            "required": true,
            "schema": {
              "title": "State",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "code",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Code"
            }
          },
          {
            "in": "query",
            "name": "error",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Error"
            }
          }
        ],
        "responses": {
          "303": {
            "description": "Redirect to the browser OAuth callback page"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {},
          {
            "CookieSession": []
          }
        ],
        "summary": "Complete Oauth Callback",
        "tags": ["identity"]
      }
    },
    "/auth/oauth/{provider}/start": {
      "get": {
        "operationId": "start_oauth_auth_oauth__provider__start_get",
        "parameters": [
          {
            "in": "path",
            "name": "provider",
            "required": true,
            "schema": {
              "title": "Provider",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "return_to",
            "required": false,
            "schema": {
              "default": "/aigc",
              "maxLength": 2048,
              "title": "Return To",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "flow",
            "required": false,
            "schema": {
              "default": "login",
              "enum": ["login", "connect"],
              "title": "Flow",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_OAuthStart_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {},
          {
            "CookieSession": []
          }
        ],
        "summary": "Start Oauth",
        "tags": ["identity"]
      }
    },
...
    "/auth/verification/complete": {
      "post": {
        "operationId": "complete_verification_login_auth_verification_complete_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerificationCompleteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_IdentitySession_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Complete Verification Login",
        "tags": ["identity"]
      }
    },
    "/auth/verification/request": {
      "post": {
        "operationId": "request_verification_code_auth_verification_request_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerificationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_VerificationChallenge_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Request Verification Code",
        "tags": ["identity"]
      }
    },
...
  }
}